+++ Sources/ShimmerKit/Core/ShimmerConfig.swift
@@ -10,6 +10,7 @@ import SwiftUI
 
 public struct ShimmerConfig: @unchecked Sendable {
     public var gradient: Gradient
+    public var textGradient: Gradient?
     public var skeletonColor: Color
     public var speed: Double
     public var angle: Angle
@@ -23,6 +24,7 @@ public struct ShimmerConfig: @unchecked Sendable {
             Color.white.opacity(0.35),
             .clear
         ]),
+        textGradient: Gradient? = nil,
         skeletonColor: Color = Color.gray.opacity(0.25),
         speed: Double = 1.2,
         angle: Angle = .degrees(20),
@@ -31,6 +33,7 @@ public struct ShimmerConfig: @unchecked Sendable {
         useLayoutProtocolIntegration: Bool = false
     ) {
         self.gradient = gradient
+        self.textGradient = textGradient
         self.skeletonColor = skeletonColor
         self.speed = speed
         self.angle = angle
@@ -41,6 +44,7 @@ public struct ShimmerConfig: @unchecked Sendable {
     
     public init(
         shimmerColor: Color,
+        textGradient: Gradient? = nil,
         skeletonColor: Color = Color.gray.opacity(0.25),
         shimmerOpacity: Double = 0.35,
         speed: Double = 1.2,
@@ -54,6 +58,7 @@ public struct ShimmerConfig: @unchecked Sendable {
             shimmerColor.opacity(shimmerOpacity),
             .clear
         ])
+        self.textGradient = textGradient
         self.skeletonColor = skeletonColor
         self.speed = speed
         self.angle = angle